Extracting the text box into its own control so that it can be used in the search bar as well. It's unfortunate that it seems like you either need to duplicate code or axaml, but I've taken the axaml duplication for now and will think about changing it in the future if issues arise. Replacing the search boxes with the new templated control. Adjusting the windows now that the validation doesn't change the size of the components vertically.
