+++ Views/Courses/CreateOrEditCourseWindow.axaml
@@ -7,13 +7,13 @@
         xmlns:controls="using:CourseEquivalencyDesktop.Views.CustomControls"
         mc:Ignorable="d"
         d:DesignWidth="900"
-        d:DesignHeight="750"
+        d:DesignHeight="700"
         x:Class="CourseEquivalencyDesktop.Views.Courses.CreateOrEditCourseWindow"
         x:DataType="vm:CreateOrEditCourseViewModel"
         Title="{Binding WindowAndButtonText}"
         CanResize="False"
         Width="900"
-        Height="750"
+        Height="700"
         WindowStartupLocation="CenterOwner">
 
     <Design.DataContext>
