+++ Views/Universities/CreateOrEditUniversityWindow.axaml
@@ -6,13 +6,13 @@
         xmlns:controls="using:CourseEquivalencyDesktop.Views.CustomControls"
         mc:Ignorable="d"
         d:DesignWidth="800"
-        d:DesignHeight="340"
+        d:DesignHeight="300"
         x:Class="CourseEquivalencyDesktop.Views.Universities.CreateOrEditUniversityWindow"
         x:DataType="vm:CreateOrEditUniversityViewModel"
         Title="{Binding WindowAndButtonText}"
         CanResize="False"
         Width="800"
-        Height="340"
+        Height="300"
         WindowStartupLocation="CenterOwner">
 
     <Design.DataContext>
